requirements.txt
accelerate==0.28.0
diffusers==0.27.1
huggingface-hub==0.21.4
python-dotenv==1.0.0
safetensors==0.4.2
torch==2.2.1
transformers==4.38.2
